# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AEGIS QUANT AI - Quantitative Trading Research Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: symbol
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Create / record a trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade (close, adjust stops, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trades/{id}/reasons:
    get:
      operationId: getTradeReasons
      tags: [trades]
      summary: Get reasons recorded for a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade reasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeReason"

  # ── ANALYTICS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trade:
      type: object
      required: [id, symbol, side, entryPrice, quantity, status, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        strategyId:
          type: ["integer", "null"]
        strategyName:
          type: ["string", "null"]
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        quantity:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        profitLoss:
          type: ["number", "null"]
        profitPercent:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, closed, cancelled]
        aiConfidence:
          type: ["number", "null"]
        timeframe:
          type: ["string", "null"]
        entryTime:
          type: string
        exitTime:
          type: ["string", "null"]
        createdAt:
          type: string
    TradeInput:
      type: object
      required: [symbol, side, entryPrice, quantity]
      properties:
        symbol:
          type: string
        strategyId:
          type: integer
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        quantity:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        aiConfidence:
          type: number
        timeframe:
          type: string
        reasons:
          type: array
          items:
            type: object
            properties:
              reasonType:
                type: string
              reasonText:
                type: string
    TradeUpdate:
      type: object
      properties:
        exitPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        status:
          type: string
          enum: [open, closed, cancelled]
        exitTime:
          type: string
    TradeReason:
      type: object
      required: [id, tradeId, reasonType, reasonText]
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        reasonType:
          type: string
        reasonText:
          type: string

    # ── ANALYTICS ───────────────────────────────────────────────────────────────
...